refactor: 💡 modify types
