+++ packages/ui-react/src/types/index.ts
@@ -1,4 +1,6 @@
 export * from './chain';
 export * from './common';
+export * from './components';
+export * from './footer';
 export * from './network';
 export * from './withdraw';
